+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Removed pos field from the license struct
+- Refactored utils to accept external license and opening
+
 ## [0.4.1] - 2023-08-09
 
 ### Added