+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.4.2] - 2023-10-12
+## [0.5.0] - 2023-10-12
 
 ### Changed
 
@@ -85,8 +85,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#21]: https://github.com/dusk-network/citadel/issues/21
 
 <!-- VERSIONS -->
-[Unreleased]: https://github.com/dusk-network/citadel/compare/v0.4.2...HEAD
-[0.4.2]: https://github.com/dusk-network/citadel/compare/v0.4.1...v0.4.2
+[Unreleased]: https://github.com/dusk-network/citadel/compare/v0.5.0...HEAD
+[0.5.0]: https://github.com/dusk-network/citadel/compare/v0.4.1...v0.5.0
 [0.4.1]: https://github.com/dusk-network/citadel/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/dusk-network/citadel/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/dusk-network/citadel/compare/v0.2.0...v0.3.0